ensure there is no NULL char among random data
This random data is later used as part of input for ngx_crypt function,
which expects a zero-terminated string. If the string contains a zero
character, then rest of string is ignored, and the seed is effectively
truncated. Good that the seed is at the end of the string. :) I believe
it's better to exclude strings containing zero character from possible
outputs of this function, rather than truncate everything after it.